working examples 1-6. Rename
