+++ examples/02-simple-input/number.py
