+++ examples/03-sequential/maximum.py
