+++ examples/04-parallel/minimum.py
