+++ examples/05-sleep/test_lazy.py
