+++ examples/06-tensorflow/mnist.py
