--- examples/old/tensor/cowait.yml
@@ -1,8 +1,0 @@
-version: 1
-cowait:
-  image: docker.backtick.se/tensor
-
-  cluster:
-    default: docker
-    docker:
-      network: tasks
